Add Authentication

+++ src/data/model/Customer.kt
@@ -5,7 +5,7 @@ data class Customer(
      val password : String,
      val first_name : String,
      val last_name : String,
-     val phone_no : Int,
+     val phone_no : Long,
      val cust_Address : String,
      val delivery_address : String,
      val pincode : Int
